refactor[CreateInterested]
