+++ src/modules/CreateInterested/dtos/InteressadosDTO.ts
@@ -1,0 +1,6 @@
+import { ILoginDTO } from "../../../DTO/LoginDTO";
+
+export interface IinteressadosDTO{
+    login: ILoginDTO;
+    etiqueta: string;
+}
